only blog modification with authenticated  user token

+++ part4/bloglist/controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
@@ -9,7 +10,12 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const user = await User.findById(request.body.userId)
+  
+  if (!request.userId) {
+    return response.status(401).json({ error: 'Token inválido' });
+  }
+  
+  const user = request.user
 
   const blog = new Blog({
     title: request.body.title,
@@ -28,8 +34,19 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
+  
+  if (!request.userId) {
+    return response.status(401).json({ error: 'Token inválido' });
+  }
 
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+
+  if (!(user._id.toString() === blog.user.toString()) || !blog.user || !user._id) {
+    return response.status(401).json({error: "this blog doesn't belongs to currently user"})
+  }
+  
+  await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
